Bind the pixel types explicitly, and port the camera geometry call sites
Dropping `forcecast` from the bindings caught a live bug within one test
run, which is the argument for dropping it.

`cv2.remap` in `ocv_stereo_disparity` runs on a **float32** depth map.
Handed to a uint8 binding with `forcecast`, pybind would have converted it
without a word: every depth past 255 clipped, silently, in a stage whose
whole output is that map. So the resamplers and the drawing kernels now
bind the pixel types a VIAME pipeline actually carries -- uint8, uint16 for
the 16-bit pipelines, float32 for a depth or disparity stage -- and
anything else raises a TypeError instead of being cast.

That immediately surfaced `resnet_augmenter`, where the earlier resize port
has been quietly wrong since it landed: the brightness jitter produces
float64, `cv2.resize` took it as float64, and
`image_kernels.resize` was truncating it to uint8 before resampling.
`unit:object_trackers:desc_augmentation` had been passing throughout,
because nothing it asserts is sensitive to the loss. It resizes in float32
now, as the original did in float64.

The camera geometry call sites follow the bindings:

  * every `cv2.Rodrigues(x)[0]` and `a, _ = cv2.Rodrigues(x)` across five
    files, which is exact -- the binding is bit identical to cv2;
  * `ocv_stereo_disparity`'s whole rectification: `stereoRectify`,
    two `initUndistortRectifyMap`, `undistortPoints` over the pixel grid,
    and the three `remap` calls, plus its two `cvtColor` greyscale paths;
  * `ocv_optimize_stereo_cameras`'s `stereoRectify` and the two
    `undistortPoints` in its epipolar error report.

`ocv_stereo_disparity` now reaches for cv2 only for the stereo matchers
themselves -- `StereoBM`, `StereoSGBM` and the ximgproc WLS filter -- which
are real OpenCV algorithms with no VIAME equivalent yet, and are a phase of
their own.

69 of our files import cv2, down from 96. 67 kernel tests pass, up from 59.

+++ library/descriptors/torchvision/resnet_augmenter.py
@@ -46,7 +46,11 @@ def augment_region( input_image, cx, cy, csize, outsize, rot, tflux=6, sflux=0.3
     iaug[ iaug > 255 ] = 255
     iaug[ iaug < 0 ] = 0
 
-    crop = image_kernels.resize(iaug, outsize, outsize)
+    # Resized as float, which is what `cv2.resize` did with the float64 the
+    # brightness jitter above produces -- the caller casts to uint8 after.
+    # float32 rather than float64 because that is the kernel's float type,
+    # and the values are 0..255.
+    crop = image_kernels.resize(iaug.astype(np.float32), outsize, outsize)
     return crop
